Moved files to subpackages Files which are specific for Play 2.X (extensions for commands) were moved to subfolder. The reason is to keep the code organized when new extensions are added for Play 1.X or Play 2.3, which will invoke sbt directly.
