+++ src/main/resources/jenkins/plugins/play/extensions/play2/PlayClean/config.jelly
