+++ src/main/resources/jenkins/plugins/play/extensions/play2/PlayCompile/config.jelly
