+++ src/main/resources/jenkins/plugins/play/extensions/play2/PlayDist/config.jelly
