+++ src/main/resources/jenkins/plugins/play/extensions/play2/PlayPackage/config.jelly
