+++ src/main/resources/jenkins/plugins/play/extensions/play2/PlayPublish/config.jelly
