+++ src/main/resources/jenkins/plugins/play/extensions/play2/PlayTest/config.jelly
