+++ src/main/resources/jenkins/plugins/play/extensions/play2/PlayTestOnly/config.jelly
